init script: Better shutdown script (HOP-4680)
Problem: Some init scripts used killall to stop the services and this
caused the init script to quit. Problem became apparent when process
monitor tried to restart an already stopped service.
While we were updating the scripts, it was proposed to unify the
functionality in amx_init_functions.sh to avoid mass updates in the
future.

Solution: Avoid using killall for stopping the service and combine the
functionality that is required in the init scripts in a single location.

Issue: HOP-4680 - [amx] Failing to restart processes with init scripts
GitOrigin-RevId: bbe0168648eba74a2a7af3b5587600517bfb5c80