baf: update targets for datamodel documentation generation
The baf system generates the top makefile, where the "doc" target is specified.
As part of the "doc" target, we have some datamodel documentation which is being generated with amxo-cg.
By default, baf searches for all odl files and feeds them to amxo-cg for inspection.

Some plugins hold other odl files than the onces used for the own datamodel. For example, components can hold amx-processmonitor odl files.
In such case, amxo-cg will file to analyze those files, as they refer to parts of the datamodel that are unknown to amxo-cg in this local scope.

This commit specifies the odl files to be examined by amxo-cg. In general only the definition files should be inspected.

References: PPW-880

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>